changed from uint8 to double type

+++ code/timeDisc.m
@@ -2,7 +2,7 @@ function c = timeDisc( g1, g2)
 %timeDisc Computes the time discretization of the time derivative
 %   Detailed explanation goes here
 
-    c = g2-g1;
+    c = double(g2)-double(g1);
 
 end
